strands_service/pyproject.toml
[project]
name = "at-ai-editor-recommender-strands-coi"
version = "0.1.0"
description = "Strands COI checker service — conflict-of-interest detection via A2A"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "starlette>=0.42.0",
    "uvicorn>=0.35.0",
    "a2a-sdk>=0.3.25",
    "strands-agents>=0.1.0",
    "httpx>=0.28.0",
    "aioboto3>=7.0.0",
    "boto3>=1.40.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["uv_build>=0.8,<0.9"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
